py/requirements.txt
absl-py==0.10.0
protobuf==3.13.0
python-dateutil==2.8.1
six==1.15.0
